% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_files.r
\name{extract_and_rename_tifs}
\alias{extract_and_rename_tifs}
\title{Batch Extract and Rename TIFs from ZIP files}
\usage{
extract_and_rename_tifs(
  zip_dir,
  output_dir,
  file_to_extract = "odm_orthophoto/odm_orthophoto.tif"
)
}
\arguments{
\item{zip_dir}{A string path to the folder \emph{containing} all the .zip files.}

\item{output_dir}{A string path to the folder where all extracted .tif files
will be saved. This folder will be created if it doesn't exist.}

\item{file_to_extract}{A string path \emph{inside} the zip file to extract.
Example: "odm_orthophoto/odm_orthophoto.tif".}
}
\value{
This function is called for its side effects (extracting and
renaming files) and returns \code{invisible(NULL)}.
}
\description{
This function iterates through all .zip files in a specified directory,
extracts a single target .tif file from each, renames that .tif file
to match its parent .zip file, and saves it to a single output directory.
}
