% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{footprint}
\alias{footprint}
\title{Get Spatial Footprint of a Raster}
\usage{
footprint(path, target_crs = NULL)
}
\arguments{
\item{path}{Character. Path to a raster file.}

\item{target_crs}{Optional CRS to reproject into.}
}
\value{
An \code{sf} object with one polygon.
}
\description{
Get Spatial Footprint of a Raster
}
