% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read an Existing Catalog}
\usage{
read_catalog(path)
}
\arguments{
\item{path}{Path to a GeoJSON catalog.}
}
\value{
An \code{sf} object.
}
\description{
Read an Existing Catalog
}
