% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_species_name.R
\name{check_unmapped_species}
\alias{check_unmapped_species}
\title{Check for Unmapped Species Codes}
\usage{
check_unmapped_species(df, species_col)
}
\arguments{
\item{df}{A data frame containing species information}

\item{species_col}{Name of the column containing species codes (unquoted)}
}
\value{
A data frame of unrecognized codes with their frequencies
}
\description{
Identifies species codes in your data that are not recognized by the
standardization function. Useful for quality control and identifying
new codes that need to be added.
}
\examples{
example_df <- data.frame(
  species = c("CBO", "IC", "UNKNOWN", "XCODE"),
  stringsAsFactors = FALSE
)

# Check for unmapped codes before standardizing
unmapped <- check_unmapped_species(example_df, species)

# If codes are found, add them to the clean_species_names function

}
