% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binary_feature.R
\name{create_binary_features}
\alias{create_binary_features}
\title{Create Binary Features}
\usage{
create_binary_features(data, target_col, positive_classes = NULL)
}
\arguments{
\item{data}{Input dataframe}

\item{target_col}{Name of the target column}

\item{positive_classes}{Vector of positive class names (NULL = use all unique values)}
}
\value{
Dataframe with new binary columns
}
\description{
Create Binary Features for Multi-class Variable
}
